refactor: moving transformEvent, transformBookings in helper function
