+++ graphql/resolvers/booking.js
@@ -1,26 +1,6 @@
 const Event = require('../../models/event');
 const Booking = require('../../models/booking');
-const { dateToString } = require('../../helpers/date');
-
-const { user, singleEvent } = require("./helper");
-
-const transformEvent = event => {
-    return {
-        ...event._doc,
-        date: dateToString(event._doc.date),
-        creator: user.bind(this, event._doc.creator)
-    };
-}
-
-const transformBooking = booking => {
-    return {
-        ...booking._doc,
-        user: user.bind(this, booking._doc.user),
-        event: singleEvent.bind(this, booking._doc.event),
-        createdAt: dateToString(booking._doc.createdAt),
-        updatedAt: dateToString(booking._doc.updatedAt)
-    }
-}
+const { transformEvent, transformBooking } = require("./helper");
 
 module.exports = {
     bookings: async () => {
@@ -38,7 +18,7 @@ module.exports = {
         try {
             const fetchedEvent = await Event.findOne({ _id: args.eventId });
             const booking = new Booking({
-                user: '65fe26b8a230d1a9d2e32c41',
+                user: '65fe2d677edbd3468f375c83',
                 event: fetchedEvent
             });
             const result = await booking.save();
