+++ graphql/resolvers/events.js
@@ -1,16 +1,7 @@
 const Event = require('../../models/event');
 const User = require('../../models/user');
-const { dateToString } = require('../../helpers/date');
 
-const { user } = require("./helper");
-
-const transformEvent = event => {
-    return {
-        ...event._doc,
-        date: dateToString(event._doc.date),
-        creator: user.bind(this, event._doc.creator)
-    };
-}
+const { transformEvent } = require("./helper");
 
 module.exports = {
     events: async () => {
@@ -29,13 +20,13 @@ module.exports = {
             description: args.eventInput.description,
             price: +args.eventInput.price,
             date: new Date(args.eventInput.date),
-            creator: '65fe26b8a230d1a9d2e32c41'
+            creator: '65fe2d677edbd3468f375c83'
         });
         let createEvent;
         try {
             const result = await event.save();
             createEvent = transformEvent(result);
-            const creator = await User.findById('65fe26b8a230d1a9d2e32c41');
+            const creator = await User.findById('65fe2d677edbd3468f375c83');
             if (!creator) {
                 throw new Error("User not found!");
             }
